refactor(native): extract shell open commands
